+++ changelog.md
@@ -6,6 +6,7 @@ Improved banner generator:
   
 - Ability to choose whether to show division or rank as front row in the right side
 - Accept custom background image defined by `banner_background_url` in config
+- Accept custom background opacity defined by `banner_background_opacity` in config
 
 ## v2.9.4